Test: refactor, pull out adapters
