+++ test/adapter/index.js
@@ -1,0 +1,8 @@
+module.exports = [
+  require("./fs"),
+  require("./github"),
+  require("./dropbox"),
+  require("./onedrive"),
+  require("./google"),
+  // require("./webdav")
+];
